Pagination Fix and Generalization
