+++ org.confery.client/src/main/java/controller/registration/ControllerLogin.java
@@ -1,6 +1,8 @@
-package controller;
+package controller.registration;
 
+import controller.main.ControllerInterface;
 import domain.UserEntity;
+import exception.SystemException;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
@@ -28,22 +30,13 @@ import static utils.Try.runFunction;
 
 @Lazy
 @Component
-public class ControllerSignUp implements ControllerInterface, SubscriberService {
+public class ControllerLogin implements ControllerInterface, SubscriberService {
 
     @FXML
     private TextField usernameTextField;
 
     @FXML
     private TextField passwordTextField;
-
-    @FXML
-    private TextField emailTextField;
-
-    @FXML
-    private TextField displayNameTextField;
-
-    @FXML
-    private TextField confirmTextField;
 
     @FXML
     private Label errorLabel;
@@ -79,6 +72,7 @@ public class ControllerSignUp implements ControllerInterface, SubscriberService 
         runFunction(listener::addSubscriber, this).orHandle(System.out::println);
     }
 
+
     /**
      * Effect: Loads the ConferencesView.
      *
@@ -89,36 +83,38 @@ public class ControllerSignUp implements ControllerInterface, SubscriberService 
         manager.switchScene(ViewType.CONFERENCES);
     }
 
+
     /**
-     * Effect: Loads the LoginView.
+     * Effect: Loads the SignUpView.
      *
      * @implNote status: In development.
      */
     @FXML
-    void onLoginButtonClick() throws RemoteException {
-        manager.switchScene(ViewType.LOGIN);
+    void onSignUpButtonClick() throws RemoteException {
+        manager.switchScene(ViewType.SIGN_UP);
     }
 
     /**
-     * Effect: The user registers in the system with his data.
+     * Effect: The user logs in the system with account data.
      *
      * @implNote status: Unavailable at the moment.
      */
     @FXML
-    void onSignUpButtonClick() throws RemoteException {
-        String email = emailTextField.getText();
-        String displayName = displayNameTextField.getText();
+    void onLoginButtonClick() throws RemoteException, SystemException {
         String username = usernameTextField.getText();
         String password = passwordTextField.getText();
-        String confirm = confirmTextField.getText();
-        UserEntity user = service.signUpService().signUp(username, password, confirm, email, displayName);
-        this.listener.setActiveUser(user);
-        this.listener.notifyAll(new NotificationUpdate(NotificationType.SIGNAL_SIGN_UP));
-        manager.switchScene(ViewType.CONFERENCES);
+        try {
+            UserEntity user = service.loginService().login(username, password);
+            listener.setActiveUser(user);
+            listener.notifyAll(new NotificationUpdate(NotificationType.SIGNAL_LOGIN));
+            manager.switchScene(ViewType.CONFERENCES);
+        } catch (RemoteException exception) {
+            errorLabel.setText(exception.getCause().getMessage());
+        }
     }
 
     @Override
     public void update(NotificationUpdate notification) throws RemoteException {
-        System.out.print(notification);
+        System.out.print(notification.getType());
     }
 }
