+++ org.confery.client/src/main/java/pagination/PaginationBuilder.java
@@ -12,6 +12,8 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 
+import static utils.Conditional.basedOn;
+
 /**
  * Special builder for pagination manager. [based on Builder Pattern]
  * <p>
@@ -30,8 +32,8 @@ public class PaginationBuilder<T, E extends PaginationControllerItemInterface<T>
 
     private ViewType view;
     private ObservableList<T> elements;
-    private Integer rows;
-    private Integer columns;
+    private Integer rows = 2;
+    private Integer columns = 4;
     private StageManager stageManager;
     private Pagination pagination;
 
@@ -92,7 +94,6 @@ public class PaginationBuilder<T, E extends PaginationControllerItemInterface<T>
      * @param rows The pagination's number of rows.
      * @return the builder for later chaining.
      */
-    @SuppressWarnings("all")
     public PaginationBuilder setRows(Integer rows) {
         this.rows = rows;
         return this;
@@ -101,10 +102,9 @@ public class PaginationBuilder<T, E extends PaginationControllerItemInterface<T>
     /**
      * Effect: Sets the pagination's number of columns and returns the builder for later chaining.
      *
-     * @param rows The pagination's number of columns.
+     * @param columns The pagination's number of columns.
      * @return the builder for later chaining.
      */
-    @SuppressWarnings("all")
     public PaginationBuilder setColumns(Integer columns) {
         this.columns = columns;
         return this;
@@ -132,6 +132,37 @@ public class PaginationBuilder<T, E extends PaginationControllerItemInterface<T>
         return this;
     }
 
+    private Boolean checkViewTypeState() {
+        basedOn(view != null).orThrow(new IllegalStateException("Missing Pagination View!"));
+        return Boolean.TRUE;
+    }
+
+    private Boolean checkStageManagerState() {
+        basedOn(stageManager != null).orThrow(new IllegalStateException("Missing Stage Manager!"));
+        return Boolean.TRUE;
+    }
+
+    private Boolean checkPaginationState() {
+        basedOn(pagination != null).orThrow(new IllegalStateException("Missing Pagination!"));
+        return Boolean.TRUE;
+    }
+
+    private Boolean checkCurrentState() {
+        return checkViewTypeState() && checkStageManagerState() && checkPaginationState();
+     }
+
+    private Boolean checkCurrentPaginationManager(PaginationManagerInterface<T, E, U> manager) {
+        basedOn(manager != null).orThrow(new IllegalStateException("Pane not supported yet!"));
+        return Boolean.TRUE;
+    }
+
+    private Pagination setUpAndBuild(PaginationManagerInterface<T, E, U> manager) {
+        manager.setElements(this.elements);
+        manager.setStageManager(stageManager);
+        manager.setView(this.view);
+        return manager.buildPagination(pagination);
+    }
+
     /**
      * Effect: Builds the pagination.
      *
@@ -139,28 +170,9 @@ public class PaginationBuilder<T, E extends PaginationControllerItemInterface<T>
      * @return the wanted pagination.
      * @throws IllegalStateException If you forget an important element for the pagination's manager.
      */
-    public Pagination build(Class<U> paneType) throws IllegalStateException {
-        if (view == null) {
-            throw new IllegalStateException("Missing Pagination View!");
-        }
-        if (rows == null || columns == null) {
-            throw new IllegalStateException("Missing Pagination Size");
-        }
-        if (stageManager == null) {
-            throw new IllegalStateException("Missing Stage Manger");
-        }
-        if (pagination == null) {
-            throw new IllegalStateException("Missing Pagination");
-        }
-        PaginationManagerInterface<T, E, U> manager =
-                PaginationFactory.getPagination(paneType, this.rows, this.columns);
-        if (manager == null) {
-            throw new IllegalStateException("Pagination's pane is not supported yet.");
-        }
-        manager.setElements(this.elements);
-        manager.setStageManager(stageManager);
-        manager.setView(this.view);
-        return manager.buildPagination(pagination);
+    public <Y> Pagination build(Class<Y> paneType) throws IllegalStateException {
+        PaginationManagerInterface<T, E, U> manager = checkCurrentState() ?
+                PaginationFactory.getPagination(paneType, this.rows, this.columns) : null;
+        return checkCurrentPaginationManager(manager) ? setUpAndBuild(manager) : null;
     }
-
 }
