+++ org.confery.client/src/main/resources/fxml/LoginView.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.image.*?>
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
-<HBox xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.ControllerLogin">
+<HBox xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.registration.ControllerLogin">
    <children>
       <VBox alignment="CENTER" minWidth="350.0" style="-fx-background-color: white;">
          <children>
