+++ org.confery.client/src/main/resources/fxml/SignUpView.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.image.*?>
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
-<HBox xmlns="http://javafx.com/javafx/8.0.112" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.ControllerSignUp">
+<HBox xmlns="http://javafx.com/javafx/8.0.112" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.registration.ControllerSignUp">
    <children>
       <VBox alignment="CENTER" minWidth="350.0" style="-fx-background-color: white;">
          <children>
